[Impeller] move everything needed by the code gen template to core
