+++ impeller/compiler/code_gen_template.h
@@ -18,13 +18,11 @@ constexpr std::string_view kReflectionHeaderTemplate =
 {# file itself and the no-check declarations will be stripped in generated files.#}
 #include "impeller/core/buffer_view.h"                {# // nogncheck #}
 
-#include "impeller/renderer/command.h"                    {# // nogncheck #}
+#include "impeller/core/sampler.h"                    {# // nogncheck #}
 
-#include "impeller/renderer/compute_command.h"            {# // nogncheck #}
+#include "impeller/core/shader_types.h"               {# // nogncheck #}
 
-#include "impeller/renderer/sampler.h"                    {# // nogncheck #}
-
-#include "impeller/renderer/shader_types.h"               {# // nogncheck #}
+#include "impeller/core/resource_binder.h"            {# // nogncheck #}
 
 #include "impeller/core/texture.h"                    {# // nogncheck #}
