+++ impeller/compiler/reflector.cc
@@ -86,21 +86,6 @@ static std::string ExecutionModelToString(spv::ExecutionModel model) {
   }
 }
 
-static std::string ExecutionModelToCommandTypeName(
-    spv::ExecutionModel execution_model) {
-  switch (execution_model) {
-    case spv::ExecutionModel::ExecutionModelVertex:
-    case spv::ExecutionModel::ExecutionModelFragment:
-    case spv::ExecutionModel::ExecutionModelTessellationControl:
-    case spv::ExecutionModel::ExecutionModelTessellationEvaluation:
-      return "Command&";
-    case spv::ExecutionModel::ExecutionModelGLCompute:
-      return "ComputeCommand&";
-    default:
-      return "unsupported";
-  }
-}
-
 static std::string StringToShaderStage(std::string str) {
   if (str == "vertex") {
     return "ShaderStage::kVertex";
@@ -1126,7 +1111,7 @@ std::vector<Reflector::BindPrototype> Reflector::ReflectBindPrototypes(
       proto.docstring = stream.str();
     }
     proto.args.push_back(BindPrototypeArgument{
-        .type_name = ExecutionModelToCommandTypeName(execution_model),
+        .type_name = "ResourceBinder&",
         .argument_name = "command",
     });
     proto.args.push_back(BindPrototypeArgument{
@@ -1145,7 +1130,7 @@ std::vector<Reflector::BindPrototype> Reflector::ReflectBindPrototypes(
       proto.docstring = stream.str();
     }
     proto.args.push_back(BindPrototypeArgument{
-        .type_name = ExecutionModelToCommandTypeName(execution_model),
+        .type_name = "ResourceBinder&",
         .argument_name = "command",
     });
     proto.args.push_back(BindPrototypeArgument{
@@ -1164,7 +1149,7 @@ std::vector<Reflector::BindPrototype> Reflector::ReflectBindPrototypes(
       proto.docstring = stream.str();
     }
     proto.args.push_back(BindPrototypeArgument{
-        .type_name = ExecutionModelToCommandTypeName(execution_model),
+        .type_name = "ResourceBinder&",
         .argument_name = "command",
     });
     proto.args.push_back(BindPrototypeArgument{
