+++ impeller/core/BUILD.gn
@@ -24,6 +24,14 @@ impeller_component("core") {
     "platform.h",
     "range.cc",
     "range.h",
+    "resource_binder.cc",
+    "resource_binder.h",
+    "sampler.cc",
+    "sampler.h",
+    "sampler_descriptor.cc",
+    "sampler_descriptor.h",
+    "shader_types.cc",
+    "shader_types.h",
     "texture.cc",
     "texture.h",
     "texture_descriptor.cc",
