+++ impeller/core/sampler.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "impeller/renderer/sampler.h"
+#include "impeller/core/sampler.h"
 
 namespace impeller {
