+++ impeller/entity/contents/contents.h
@@ -9,10 +9,10 @@
 #include <vector>
 
 #include "flutter/fml/macros.h"
+#include "impeller/core/sampler_descriptor.h"
 #include "impeller/core/texture.h"
 #include "impeller/geometry/color.h"
 #include "impeller/geometry/rect.h"
-#include "impeller/renderer/sampler_descriptor.h"
 #include "impeller/renderer/snapshot.h"
 
 namespace impeller {
