+++ impeller/entity/contents/runtime_effect_contents.h
@@ -6,8 +6,8 @@
 #include <memory>
 #include <vector>
 
+#include "impeller/core/sampler_descriptor.h"
 #include "impeller/entity/contents/color_source_contents.h"
-#include "impeller/renderer/sampler_descriptor.h"
 #include "impeller/runtime_stage/runtime_stage.h"
 
 namespace impeller {
