+++ impeller/renderer/BUILD.gn
@@ -78,10 +78,6 @@ impeller_component("renderer") {
     "render_target.h",
     "renderer.cc",
     "renderer.h",
-    "sampler.cc",
-    "sampler.h",
-    "sampler_descriptor.cc",
-    "sampler_descriptor.h",
     "sampler_library.cc",
     "sampler_library.h",
     "shader_function.cc",
@@ -90,8 +86,6 @@ impeller_component("renderer") {
     "shader_key.h",
     "shader_library.cc",
     "shader_library.h",
-    "shader_types.cc",
-    "shader_types.h",
     "snapshot.cc",
     "snapshot.h",
     "surface.cc",
