+++ impeller/renderer/backend/gles/formats_gles.h
@@ -7,8 +7,8 @@
 #include "flutter/fml/logging.h"
 #include "flutter/fml/macros.h"
 #include "impeller/core/formats.h"
+#include "impeller/core/shader_types.h"
 #include "impeller/renderer/backend/gles/gles.h"
-#include "impeller/renderer/shader_types.h"
 
 namespace impeller {
