+++ impeller/renderer/backend/gles/sampler_library_gles.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "flutter/fml/macros.h"
-#include "impeller/renderer/sampler_descriptor.h"
+#include "impeller/core/sampler_descriptor.h"
 #include "impeller/renderer/sampler_library.h"
 
 namespace impeller {
