+++ impeller/renderer/backend/metal/blit_pass_mtl.mm
@@ -13,6 +13,7 @@
 #include "impeller/base/backend_cast.h"
 #include "impeller/core/formats.h"
 #include "impeller/core/host_buffer.h"
+#include "impeller/core/shader_types.h"
 #include "impeller/renderer/backend/metal/blit_command_mtl.h"
 #include "impeller/renderer/backend/metal/device_buffer_mtl.h"
 #include "impeller/renderer/backend/metal/formats_mtl.h"
@@ -20,7 +21,6 @@
 #include "impeller/renderer/backend/metal/sampler_mtl.h"
 #include "impeller/renderer/backend/metal/texture_mtl.h"
 #include "impeller/renderer/blit_command.h"
-#include "impeller/renderer/shader_types.h"
 
 namespace impeller {
