+++ impeller/renderer/backend/vulkan/sampler_library_vk.h
@@ -7,8 +7,8 @@
 #include "flutter/fml/macros.h"
 #include "impeller/base/backend_cast.h"
 #include "impeller/base/comparable.h"
+#include "impeller/core/sampler_descriptor.h"
 #include "impeller/renderer/backend/vulkan/vk.h"
-#include "impeller/renderer/sampler_descriptor.h"
 #include "impeller/renderer/sampler_library.h"
 
 namespace impeller {
