+++ impeller/renderer/backend/vulkan/sampler_vk.h
@@ -6,9 +6,9 @@
 
 #include "flutter/fml/macros.h"
 #include "impeller/base/backend_cast.h"
+#include "impeller/core/sampler.h"
 #include "impeller/renderer/backend/vulkan/shared_object_vk.h"
 #include "impeller/renderer/backend/vulkan/vk.h"
-#include "impeller/renderer/sampler.h"
 #include "vulkan/vulkan_handles.hpp"
 
 namespace impeller {
