+++ impeller/renderer/compute_command.h
@@ -13,14 +13,15 @@
 #include "flutter/fml/macros.h"
 #include "impeller/core/buffer_view.h"
 #include "impeller/core/formats.h"
+#include "impeller/core/resource_binder.h"
+#include "impeller/core/sampler.h"
+#include "impeller/core/shader_types.h"
 #include "impeller/core/texture.h"
 #include "impeller/core/vertex_buffer.h"
 #include "impeller/geometry/rect.h"
 #include "impeller/renderer/command.h"
 #include "impeller/renderer/compute_pipeline_descriptor.h"
 #include "impeller/renderer/pipeline.h"
-#include "impeller/renderer/sampler.h"
-#include "impeller/renderer/shader_types.h"
 #include "impeller/renderer/vertex_buffer_builder.h"
 #include "impeller/tessellator/tessellator.h"
 
@@ -39,7 +40,7 @@ namespace impeller {
 ///             views into buffers managed by other allocators and resource
 ///             managers.
 ///
-struct ComputeCommand {
+struct ComputeCommand : public ResourceBinder {
   //----------------------------------------------------------------------------
   /// The pipeline to use for this command.
   ///
@@ -54,26 +55,30 @@ struct ComputeCommand {
   ///
   std::string label;
 
+  // |ResourceBinder|
   bool BindResource(ShaderStage stage,
                     const ShaderUniformSlot& slot,
                     const ShaderMetadata& metadata,
-                    const BufferView& view);
+                    const BufferView& view) override;
 
+  // |ResourceBinder|
   bool BindResource(ShaderStage stage,
                     const SampledImageSlot& slot,
                     const ShaderMetadata& metadata,
-                    const std::shared_ptr<const Texture>& texture);
+                    const std::shared_ptr<const Texture>& texture) override;
 
+  // |ResourceBinder|
   bool BindResource(ShaderStage stage,
                     const SampledImageSlot& slot,
                     const ShaderMetadata& metadata,
-                    const std::shared_ptr<const Sampler>& sampler);
+                    const std::shared_ptr<const Sampler>& sampler) override;
 
+  // |ResourceBinder|
   bool BindResource(ShaderStage stage,
                     const SampledImageSlot& slot,
                     const ShaderMetadata& metadata,
                     const std::shared_ptr<const Texture>& texture,
-                    const std::shared_ptr<const Sampler>& sampler);
+                    const std::shared_ptr<const Sampler>& sampler) override;
 
   constexpr operator bool() const { return pipeline && pipeline->IsValid(); }
 };
