+++ impeller/renderer/compute_pipeline_descriptor.h
@@ -17,7 +17,7 @@
 #include "flutter/fml/macros.h"
 #include "impeller/base/comparable.h"
 #include "impeller/core/formats.h"
-#include "impeller/renderer/shader_types.h"
+#include "impeller/core/shader_types.h"
 #include "impeller/tessellator/tessellator.h"
 
 namespace impeller {
