+++ impeller/renderer/vertex_descriptor.h
@@ -8,7 +8,7 @@
 
 #include "flutter/fml/macros.h"
 #include "impeller/base/comparable.h"
-#include "impeller/renderer/shader_types.h"
+#include "impeller/core/shader_types.h"
 
 namespace impeller {
